sceneを分けて、efigame.hを修正
